Further clean up. Add copyright statements.
